refactor(auth): inline UserId into auth-types.ts, delete ids.ts
auth/src/ids.ts had two symbols (UserId, asUserId) with one src consumer
inside the package: auth-types.ts (where AuthUser.id and PersistedAuth.userId
are typed UserId). Two more were node-side. The dedicated file dated back to
when OwnerId/asOwnerId/TEAM_OWNER_ID lived there too; with those moved to
@epicenter/constants/identity in 82fa3afd4, the file no longer earned the
split.

UserId and asUserId now live at the top of auth-types.ts, immediately above
the AuthUser schema that uses the brand. node/machine-auth.ts and its test
import from auth-types.js. The public re-export through index.ts stays:
UserId is reachable via AuthUser.id and PersistedAuth.userId in consumer
code, and the named handle lets callers assert or brand without a deep import.
