+++ packages/auth/src/index.ts
@@ -2,9 +2,11 @@ export { type AuthClient, type AuthState } from './auth-contract.js';
 export * from './auth-errors.js';
 export {
 	ApiSessionResponse,
+	asUserId,
 	AuthUser,
 	type OAuthTokenGrant,
 	PersistedAuth,
+	UserId,
 } from './auth-types.js';
 export {
 	type AuthFetch,
@@ -14,4 +16,3 @@ export {
 	type PersistedAuthStorage,
 } from './create-oauth-app-auth.js';
 export { createTestAuth } from './create-test-auth.js';
-export { asUserId, UserId } from './ids.js';
