+++ packages/auth/src/node/machine-auth.test.ts
@@ -16,7 +16,7 @@ import * as path from 'node:path';
 import { expectErr, expectOk } from 'wellcrafted/testing';
 import type { PersistedAuth } from '../auth-types.js';
 import { asOwnerId } from '@epicenter/constants/identity';
-import { asUserId } from '../ids.js';
+import { asUserId } from '../auth-types.js';
 import type { AuthFetch } from '../create-oauth-app-auth.js';
 import {
 	createMachineAuthClient,
